Refactor: Move Elasticsearch services to application layer
- Moved ElasticsearchSyncService and ProjectSearchService from infrastructure to application layer
- Updated all imports in controllers, event listeners, and tests
- Services should reside in application layer, not infrastructure/persistence
- Applied spotless formatting
- Verified app builds and runs successfully
