+++ src/main/java/com/iyte_yazilim/proje_pazari/presentation/controllers/SearchController.java
@@ -1,7 +1,7 @@
 package com.iyte_yazilim.proje_pazari.presentation.controllers;
 
+import com.iyte_yazilim.proje_pazari.application.services.ProjectSearchService;
 import com.iyte_yazilim.proje_pazari.domain.models.ApiResponse;
-import com.iyte_yazilim.proje_pazari.infrastructure.persistence.elasticsearch.services.ProjectSearchService;
 import com.iyte_yazilim.proje_pazari.infrastructure.persistence.models.ProjectDocument;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
